requirements.txt
flask==2.0.3
Werkzeug==2.2.2
mysql-connector-python